code/lesson2_openai_agents_sdk/requirements.txt
openai~=1.102.0
openai-agents~=0.2.9
python-dotenv~=1.1.1